Adding a test to make sure party selection appears before stateless when you don't have the right party to instantiate
